Release - fix error handling
