+++ modules/diff.js
@@ -3,46 +3,11 @@ const NetlifyAPI = require('netlify')
 const { gCalendarGetEvents } = require('../google-api/googleapis')
 const { parseCommits } = require('@eqworks/release')
 
-const { userInGroup, invokeSlackWorker, errMsg, getSpecificGroupIds } = require('./util')
+const { SERVICES, CLIENTS } = require('./products')
+const { userInGroup, invokeSlackWorker, errMsg } = require('./util')
 
 const { GITHUB_TOKEN, COMMIT_LIMIT = 5, NETLIFY_TOKEN, DEPLOYED = false } = process.env
 
-
-// TODO: only supports 2-stage comparison for now
-const SERVICES = {
-  overseer: {
-    baseURL: 'httsp://api.eqworks.io',
-    stages: ['dev', 'beta'],
-    key: 'OVERSEER_VER',
-    groups: getSpecificGroupIds(['flashteam', 'overseerteam', 'overlordteam'])
-  },
-  firstorder: {
-    baseURL: 'https://api.locus.place',
-    stages: ['dev', 'prod'],
-    key: 'API_VER',
-    groups: getSpecificGroupIds(['firstorderteam', 'snoketeam'])
-  },
-  keywarden: {
-    baseURL: 'https://auth.eqworks.io',
-    stages: ['dev', 'prod'],
-    key: 'KEYWARDEN_VER',
-    groups: getSpecificGroupIds(['firstorderteam', 'snoketeam', 'flashteam', 'overseerteam', 'overlordteam'])
-  },
-}
-const CLIENTS = {
-  overlord: {
-    siteId: 'overlord.eqworks.io',
-    stages: ['master', 'prod'],
-    head: 'master',
-    groups: getSpecificGroupIds(['flashteam', 'overseerteam', 'overlordteam'])
-  },
-  snoke: {
-    siteId: 'console.locus.place',
-    stages: ['dev', 'prod'],
-    head: 'master',
-    groups: getSpecificGroupIds(['firstorderteam', 'snoketeam'])
-  },
-}
 
 const mayBreak = (message) => ['break', 'incompat'].some((p) => message.toLowerCase().includes(p))
